+++ block2/src/lib.rs
@@ -83,7 +83,7 @@
 #![warn(clippy::cargo)]
 #![warn(clippy::ptr_as_ptr)]
 // Update in Cargo.toml as well.
-#![doc(html_root_url = "https://docs.rs/block2/0.2.0-alpha.4")]
+#![doc(html_root_url = "https://docs.rs/block2/0.2.0-alpha.5")]
 
 extern crate alloc;
 extern crate std;