+++ block2/CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Implemented `Debug` for `Block`, `ConcreteBlock`, `RcBlock` and
   `GlobalBlock`.
 
+### Changed
+* **BREAKING**: Updated `objc2-encode` to `v2.0.0-pre.1`.
+
 
 ## 0.2.0-alpha.4 - 2022-06-13